+++ src/interfaces/collections/messages/collections-query.ts
@@ -2,10 +2,10 @@ import type { AuthCreateOptions, Authorizable, AuthVerificationResult } from '..
 import type { CollectionsQueryDescriptor, CollectionsQuerySchema } from '../types';
 import { DIDResolver } from '../../../did/did-resolver';
 import { Message } from '../../../core/message';
+import { MessageStore } from '../../../store/message-store';
 import { removeUndefinedProperties } from '../../../utils/object';
 import { sign, verifyAuth } from '../../../core/auth';
 import { validate } from '../../../validation/validator';
-import { MessageStore } from '../../../store/message-store';
 
 type CollectionsQueryOptions = AuthCreateOptions & {
   target: string;
